[FIX]useGetStore fix
커스텀훅에서 쿼리를 바로 받도록 수정
